fix: move selectors utils to separate module
