+++ cli/src/cmd/forge/fourbyte.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 use clap::Parser;
 use ethers::prelude::artifacts::output_selection::ContractOutputSelection;
-use foundry_utils::{import_selectors, SelectorImportData};
+use foundry_utils::selectors::{import_selectors, SelectorImportData};
 
 #[derive(Debug, Clone, Parser)]
 pub struct UploadSelectorsArgs {
